Added the same analysis of occurrence of two words in articles but this time with viscos and solut
By curiosity and intuition, I wanted to know if the majority of the articles about viscosity was also about polymer solution. More than half of articles about viscosity are mentioning solution.

I've done a fast check with a copy paste (bad :-1: ) but I think it might be interesting to define a function with this type of analysis.