Adding nb_citations as factor to the score
If the article cites fewer than 10 times than it loses a point.

I also started the section to add a point it there is a common author’s name   as a writer of the article. This section is not working yet because I need to convert the most common author’s list of tuple into list of listed names.