Testing plotting
The PlotMicrofibByYear.svg still do not show the first number of the year BUT it's shown in Spyder Plots

By curiosity, did the same analysis with "scaffold". It returns the EXACT SAME bar. To investigate.

The histogram of the number of articles in function of the publication year does not show anymore in Spyder. I didn't realize when it stopped. I want to use Matplotlib to save it in a folder.