Properly converted article by year pandas's plot to matplotlib's plot
I change to matplotlib because I wanted to save the figure in a folder.

I put the year into integers to remove the decimal then I put it into string to remove the empty bar. Otherwise matplotlib take the range of values between the min and max values inside the list. In comparison, pandas only create a bar for specified x values even if they are floats.

Also, the issue of the year being cut was fixed by increasing the bottom of the figure.

I put in comments the other section that create plots because it keeps changing code in the .svg files.