Replacing plot_word_by_year in WordAnalysis by the test one
I replaced in WordAnalysis the plot_word_by_year function by the supposedly working one. I did a quick test and I saw that was not an integer yet I thought that I had converted it.

So I change the third test to add a float and converted the pub year of the return dataframe into int then into str.

Also, to simplify the code I removed the input plot_title.