Changed the database : uncompleted exportation cause unknown.
I compared my database with my colleague and mine was incomplete.

We reproduced the exact manipulation and I still get a blank cell where my colleague get values. Something bothering me was that when I was selecting my "not supposed empty cell" Excel was writing in the bottom that the cell was not empty (counting it) (police was not in white I checked).

I used Firefox or Brave as a browser. For my investigation I only used Firefox. I reproduced the same step with small changes, and the result was still the same. Once I wait a little bit longer before clicking OK to open my export files and the screen change his colour (brighter). Then the database was more completed. I tried to reproduce this bug by refreshing the page, closing browser, doing it fast or slow and I could not reproduce the bug. From now on I always get the more completed version of the database.