Data mined third and fourth article
The fourth article (5.yaml) had  various material analysis so the .yaml template is not really good for it. I am committing it because I organized some data by materials and I will merge all of it in a min max range.

REMARK : I tried to put all the solute then all their solvent in the same order.