Trying to find a way in python to datamined with html and regex
I created a list of URL to test accessing content from it but they are protected by a firewall dans it would need to much time for me to learn how to work with it.

So now I'm saving the web page in a .html file and then I read it. But I got some issue because when I'm saving and opening the saved file an error appear and I only got the first part of the page. It seem that it might be solve by opening the file in VSCodium and then saving it to a new html file

Also, I choose 5 articles to begin with the exploration with regex.
# list of URL to experiment the coding (from articles already data mined)
# 4.yaml → 20 → melt
# 5.yaml → 145 → multiple materials (too much data; min max global) + immersion
# 12.yaml → 165 → dia fibre x±y
# 13.yaml → 166 → not about RJS, surprise!!
# 24.yaml → 52 → melt