Iterated in all saved html file to check if we properly find the speed range
https://www.codegrepper.com/code-examples/python/how+to+loop+through+list+of+files+in+a+folder+in+python

I have changed the regex because I have found many variants in the writing of the spinneret speed as per rpm, RPM, r min-1. Also, I saw that I did get some other rpm that wasn't the one of the spinneret as the feeder/extruder so I put a minimum of an order of two for it. Even though I still got a mistake because I do get a collector rpm as data for my spinneret but I think I will live with it (005.html : "The speed of the spinning reservoir ranged from 15 000 to 65 000 RPM, while collector speed was 350 RPM.").

To add, I got some deviation because articles often mention the range that the spinneret can go but scientists use smaller range to do their experiment. I will also live with that.

In the cleaning process, I add a step to remove space and commas because some articles write 10 000 or 10,000 or 10000.

I did not take in charge the scenario that somebody wrote a range like this 2000-4000. Usually I'm lucky and they mention again specific rpm for some test or result.

I put a min speed to begin with that I kind of know we will obtain a smaller one (min0 = 300000000). Same for the max value.

I use if isinstance while looping for min and max because by removing empty groups in each regex match. I sometimes get only one integer and we can't iterate if there is only one.

Sometime the datamining does not find anything as in 013.html (not about RJS so it is not surprising that it did not mention rpm but in another view it could have been bad if we did get rpm that it is of no concern about our studied process..... to be thought) [it not speed:]

I named the html file with three numbers because it gave the proper sorting for getting the test as per the SortedArcticlesByScores. Note that in github it also works like that but indexing can't start with a 0 exempt for zero.
----------------------------------------------------------------------------------------------
REMARK :
60% of success for the speed range as per manually mined data (.yaml) according to a sample of 5 precisely chosen articles.
3 deviations are somewhat not too bad I think. They are often du to the authors mentioning the specification of the machine while using the full range of it for their experiments
1 complete mistake : speed of collector and not of spinneret.