Regex change for 024.html (fiber diameter range)
Non-breaking space (espace insécable)
dia ... fiber (subject reversed + words between)
about ≈ ...

When there is only one number (not a range), it is better not to put the units as zero or one because it creates an error by continuing to another number not related.

test1,  :+1:
test2, :+1:  do not get random fiber diameter
test3,  :+1:  do not get another dimension
test4,  :+1:  do not get another dimension
test5,  :-1:  do not get table's values
test6,  :-1:  do not get paranthesis's
test7,  :+1:
test8,  :+1:

Compared to the manually data mined for this article, the min and max are found even with the tests that didn't work.