Added nozzle and temperature range & used materials
Weakness of nozzles diameter :
- radius are not considered
- many used nouns : nozzle/syringe/needle tip/hole/orifice
- syringe gauge (conversion) re not considered

materials_list :
is not complete nor perfectly clean
is based on ASTM-D3853-12 and ASTM-D1600-18 (side code not in repo)

SmartStoplist :
same as in WordAnalysis could have referred to his path but I decided to keep both folder independently

used materials :
works but need amelioration. Some article used numeration and those are easily detected in regex so I removed their introduction and references because its mention other materials that are not studied in the paper.

All the scripts are tested with the same sample of 5 articles as speed and fiber diameter range.