Fixing bug due to the variety of scenario
Sometime regex to flush intro match more than once and it causes a multiple splitting. To keep the core of the article I use max(str in list). This way the code is more often accurate. I also add the flushing intro to fiber diameter because many articles mention most common fiber range in their intro and those data are falsifying ours.

Regex [] is corrected and \s+ is added to make the code more successful in finding match. I think there is a tricky part while saving the articles to html. Maybe when I saved with different browsers it has an impact. For sure I know that I saved some html on brave for my test and there were so many less newlines (\n) then when I have saved it with FireFox. \s incorporates space and newline ....etc.

There are so many types of space used in articles. I needed to add some in re.split

While processing ±, reindexing + removing a line was causing line calling dysfunction

To convert the data, I have some regex that gets three digits before getting a unit so I needed to add a part to detect if one float is followed by 2 other floats.